wishlist copilot 수정 완료
