+++ app/(afterLogin)/wishlist/page.tsx
@@ -3,7 +3,7 @@
 import {useUserWishlist} from "@/hooks/afterLogin/wishlist/useWishlist";
 import Wishlist from "@/components/pages/afterLogin/wishlist/Wishlist";
 const WishlistPage = () => {
-  const {wishlist, loading} = useUserWishlist();
+  const {loading} = useUserWishlist();
 
   if (loading) return <div>찜 한 내역을 불러오는 중..</div>;
