+++ hooks/afterLogin/users/useUserPlanCourse.ts
@@ -8,6 +8,7 @@ export const useUserPlanCourses = () => {
   const [activeTab, setActiveTab] = useState<"scheduled" | "completed">(
     "scheduled"
   );
+  const [error, setError] = useState<Error | null>(null);
 
   useEffect(() => {
     setLoading(true);
@@ -17,7 +18,9 @@ export const useUserPlanCourses = () => {
       userCourseApi.getCoursesByStatus("DONE"),
     ])
       .then(([scheduled, done]) => setAllCourses([...scheduled, ...done]))
-      .catch(() => alert("훈련과정 내역을 불러오는 데 실패했습니다."))
+      .catch((err) => {
+        setError(new Error("훈련과정 내역을 불러오는 데 실패했습니다."));
+      })
       .finally(() => setLoading(false));
   }, []);
 
@@ -34,5 +37,6 @@ export const useUserPlanCourses = () => {
     courses,
     activeTab,
     setActiveTab,
+    error,
   };
 };
